Move the remaining desktop-only tests into root e2e/tests folder
