+++ test/e2e/tests/deployment-analysis-signin.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
-import { navigateToAppointmentAndSignIn } from '../../utils/utils';
-import { PLAYWRIGHT_TAG_DEPLOYMENT_ANALYSIS } from '../../const/constants';
+import { navigateToAppointmentAndSignIn } from '../utils/utils';
+import { PLAYWRIGHT_TAG_DEPLOYMENT_ANALYSIS } from '../const/constants';
 
 /**
  * Deployment-analysis freight-verification smoke (@deployment-analysis).
@@ -13,7 +13,7 @@ import { PLAYWRIGHT_TAG_DEPLOYMENT_ANALYSIS } from '../../const/constants';
  * SAFE to run on every freight: sign-in only -- no booking, no email, no availability/
  * settings mutation (it does not use auth.desktop.setup or @stage-sanity).
  *
- * Env consumed (via ../../const/constants -> process.env, i.e. test/e2e .env):
+ * Env consumed (via ../const/constants -> process.env, i.e. test/e2e .env):
  *   APPT_URL          public edge, e.g. https://appointment.tb-dev.thunderbird.dev/ (trailing slash).
  *   APPT_TARGET_ENV   any value other than the literal `dev` (drives the OIDC path).
  *   TB_ACCTS_EMAIL / TB_ACCTS_PWORD   tbpro test-user creds (TBAcctsPage.signIn).
